Refactoring: merged UI and UIUtil
