+++ src/main/java/de/uni_hannover/sra/minimax_simulator/ui/gui/util/MemoryExportWorker.java
@@ -4,7 +4,7 @@ import de.uni_hannover.sra.minimax_simulator.io.IOUtils;
 import de.uni_hannover.sra.minimax_simulator.model.machine.base.memory.MachineMemory;
 import de.uni_hannover.sra.minimax_simulator.model.machine.base.memory.MemoryState;
 import de.uni_hannover.sra.minimax_simulator.resources.TextResource;
-import de.uni_hannover.sra.minimax_simulator.ui.UI;
+import de.uni_hannover.sra.minimax_simulator.ui.UIUtil;
 import de.uni_hannover.sra.minimax_simulator.ui.gui.components.dialogs.FXDialog;
 import javafx.scene.control.Alert.AlertType;
 
@@ -71,7 +71,7 @@ public class MemoryExportWorker implements Runnable {
 			fos = new FileOutputStream(file);
 			doExport(fos);
 		} catch (IOException ioe) {
-			UI.invokeInFAT(new Runnable() {
+			UIUtil.invokeInFAT(new Runnable() {
 				@Override
 				public void run() {
 					FXDialog fnw = new FXDialog(AlertType.ERROR, res.get("memory.export.error"), res.format("memory.export.write-error", file.getPath()));
