+++ src/main/java/de/uni_hannover/sra/minimax_simulator/ui/gui/util/MemoryImportWorker.java
@@ -5,7 +5,7 @@ import de.uni_hannover.sra.minimax_simulator.io.IOUtils;
 import de.uni_hannover.sra.minimax_simulator.model.machine.base.memory.MachineMemory;
 import de.uni_hannover.sra.minimax_simulator.model.machine.base.memory.MemoryState;
 import de.uni_hannover.sra.minimax_simulator.resources.TextResource;
-import de.uni_hannover.sra.minimax_simulator.ui.UI;
+import de.uni_hannover.sra.minimax_simulator.ui.UIUtil;
 import de.uni_hannover.sra.minimax_simulator.ui.gui.components.dialogs.FXDialog;
 import de.uni_hannover.sra.minimax_simulator.util.Util;
 import javafx.scene.control.Alert.AlertType;
@@ -83,7 +83,7 @@ public class MemoryImportWorker implements Runnable {
 			memory.setNotifiesListeners(false);
 			doImport(fis);
 		} catch (IOException ioe) {
-			UI.invokeInFAT(new Runnable() {
+			UIUtil.invokeInFAT(new Runnable() {
 				@Override
 				public void run() {
 					FXDialog fne = new FXDialog(AlertType.ERROR, res.get("memory.import.error"), res.format("memory.import.file-not-existing", file.getPath()));
@@ -152,7 +152,7 @@ public class MemoryImportWorker implements Runnable {
 			int maxAddress = memory.getMaxAddress();
 			int width = memory.getAddressWidth();
 
-			UI.invokeInFAT(new Runnable() {
+			UIUtil.invokeInFAT(new Runnable() {
 				@Override
 				public void run() {
 					FXDialog trunc = new FXDialog(AlertType.WARNING, res.get("memory.import.warning"), res.format("memory.import.bytes-truncated", Util.toHex(maxAddress, width, true), truncated));
